# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: string
        type: array
      slug:
        type: string
      store_id:
        type: string
      tenant_id:
        type: string
    required:
    - name
    - slug
    type: object
...
  dto.PermissionResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionResponse'
        type: array
      slug:
        type: string
      store_id:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.StockMovementCreateRequest:
    properties:
      note:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      reference_id:
        type: string
      reference_type:
        description: 'contoh: "transaction", "order", "purchase"'
        type: string
      tenant_id:
        type: string
      type:
        description: validasi hanya "in" atau "out"
        enum:
        - in
        - out
        type: string
    required:
    - product_id
    - quantity
    - reference_type
    - type
    type: object
  dto.StockMovementResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      note:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      reference_id:
        type: string
      reference_type:
        type: string
      type:
        type: string
    type: object
...
  dto.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: string
        type: array
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Point of Sale (POS) application with multi-tenant support.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: POS Backend API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Retrieve list of roles (optionally filtered by tenant_id)
      parameters:
      - description: Tenant ID
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Role
  /roles/:
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Create Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Role
  /roles/{id}:
    delete:
      description: Delete a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Role
    get:
      description: Retrieve a single role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update an existing role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Role
  /roles/filter:
    get:
      description: Retrieve roles using name filter, pagination and tenant filter
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Tenant ID
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get roles with filter and pagination
      tags:
      - Role
  /stock-movements:
    get:
      description: Retrieve list of all stock movements
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StockMovementResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all stock movements
      tags:
      - StockMovement
  /stock-movements/:
    post:
      consumes:
      - application/json
      description: Create a new stock movement record
      parameters:
      - description: Create Stock Movement Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StockMovementCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StockMovementResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create stock movement
      tags:
      - StockMovement
  /stock-movements/product/{id}:
    get:
      description: Retrieve stock movements for a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StockMovementResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get stock movements by product
      tags:
      - StockMovement
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"